// frontend/src/lib/api/generated/models/AttendeeAppliedEvent.ts
/**
 * Roominar
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import type { AttendeeSurveyResponseResultItem } from '../models/AttendeeSurveyResponseResultItem';
import type { AttendeeTicketTransaction } from '../models/AttendeeTicketTransaction';
import { HttpFile } from '../http/http';

export class AttendeeAppliedEvent {
    'id': number;
    'applicationId': number;
    'appliedAt': Date;
    'name': string;
    'coverImageUrl'?: string | null;
    'startAt': Date;
    'endAt': Date;
    'applicationStartAt': Date;
    'applicationEndAt': Date;
    'checkInId'?: number | null;
    'checkedInAt'?: Date | null;
    'surveyResponseResults'?: Array<AttendeeSurveyResponseResultItem>;
    'isBookmarked'?: boolean | null;
    'transactionHistories'?: Array<AttendeeTicketTransaction>;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "number",
            "format": ""
        },
        {
            "name": "applicationId",
            "baseName": "application_id",
            "type": "number",
            "format": ""
        },
        {
            "name": "appliedAt",
            "baseName": "applied_at",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "string",
            "format": ""
        },
        {
            "name": "coverImageUrl",
            "baseName": "cover_image_url",
            "type": "string",
            "format": ""
        },
        {
            "name": "startAt",
            "baseName": "start_at",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "endAt",
            "baseName": "end_at",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "applicationStartAt",
            "baseName": "application_start_at",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "applicationEndAt",
            "baseName": "application_end_at",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "checkInId",
            "baseName": "check_in_id",
            "type": "number",
            "format": ""
        },
        {
            "name": "checkedInAt",
            "baseName": "checked_in_at",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "surveyResponseResults",
            "baseName": "survey_response_results",
            "type": "Array<AttendeeSurveyResponseResultItem>",
            "format": ""
        },
        {
            "name": "isBookmarked",
            "baseName": "is_bookmarked",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "transactionHistories",
            "baseName": "transaction_histories",
            "type": "Array<AttendeeTicketTransaction>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return AttendeeAppliedEvent.attributeTypeMap;
    }

    public constructor() {
    }
}